refactor namespace to split the scope of holiday from calendar
