+++ Delsoft.Calendars.Belgian/BelgianCalendar.cs
@@ -1,4 +1,6 @@
-﻿namespace Delsoft.Calendars.Belgian;
+﻿using Delsoft.Calendars.Belgian.Holidays;
+
+namespace Delsoft.Calendars.Belgian;
 
 public class BelgianCalendar : BaseCalendar<BelgianHolidaysCalendar>
 {
